Rename pathname to subfolders
