+++ packages/cli/src/lib/setups/i18n/templates/subfolders.ts
